requirements.txt
pyaudio
pygame
pystray
Pillow
